Stop and restart all applications on deployment
React builds take around half a gig of memory when
running, and my projects now take up over a half as
well. Given that my host only has 1Gb of memory, I
either need to upgrade the host, or find a way to
reduce the amount of memory used in this process.

 To allow me to continue building projects without
having to upgrade my host, I'll stop all of my projects
before building, which should free up enough memory
to get the build passing. The only downside is that
all of my projects will periodically go offline for
a short period of time.